refactor(mcp): fold the duplicated widenedDomainsFromHistory into lib/mcp-core
The admin and partner slicers are deliberately parallel modules over different
domain unions — that is two registries, not duplication. But the rule for
reading a widening back out of conversation history is not surface-specific:
it is a fact about the message-part shape the AI SDK emits, and it had been
copied byte-for-byte into both files. Two copies of a parser is two places to
fix when the transport changes its part shape, and only one of them will be
remembered.

Now one generic implementation in lib/mcp-core, parameterised by the load-tool
name and the selectable domain list — the only two things that actually
differed. Both surfaces keep their existing exported signature, so no call site
moves and both suites (112 tests) pass untouched, which is what makes this
provably behaviour-preserving rather than merely plausible.

Imported from the submodule rather than the barrel on purpose: both slicers are
otherwise type-only consumers of mcp-core, and pulling the barrel in for one
function would drag dispatch/proxy/server into every context that slices a
registry.
